move all test into one test class and check if money event is given correctly
