+++ src/main/java/org/matsim/run/policies/KlimaTaler.java
@@ -106,9 +106,8 @@ public class KlimaTaler implements PersonDepartureEventHandler, PersonArrivalEve
     public void notifyAfterMobsim(AfterMobsimEvent afterMobsimEvent) {
         for (Map.Entry<Id<Person>, Double> idDoubleEntry : distanceTravelledWalk.entrySet()) {
             Id<Person> person = idDoubleEntry.getKey();
-            double emissionsSaved = idDoubleEntry.getValue() * 1;
-            double klimaTaler = emissionsSaved / 1 * 1;
-            System.out.println("WalkDistance" + idDoubleEntry.getValue());
+            double emissionsSaved = idDoubleEntry.getValue() * 0.176;
+            double klimaTaler = emissionsSaved / 500 * 10;
             afterMobsimEvent.getServices().getEvents().processEvent(new PersonMoneyEvent(Time.MIDNIGHT, person, klimaTaler, "klimaTalerForWalk", null, null));
         }
 
@@ -116,15 +115,15 @@ public class KlimaTaler implements PersonDepartureEventHandler, PersonArrivalEve
             Id<Person> person = idDoubleEntry.getKey();
             System.out.println(idDoubleEntry.getKey());
             System.out.println(distanceTravelledBike.entrySet());
-            double emissionsSaved = idDoubleEntry.getValue() * 1;
-            double klimaTaler = emissionsSaved / 1 * 1;
+            double emissionsSaved = idDoubleEntry.getValue() * 0.176;
+            double klimaTaler = emissionsSaved / 500 * 10;
             afterMobsimEvent.getServices().getEvents().processEvent(new PersonMoneyEvent(Time.MIDNIGHT, person, klimaTaler, "klimaTalerForBike", null, null));
         }
 
         for (Map.Entry<Id<Person>, Double> idDoubleEntry : distanceTravelledPt.entrySet()) {
             Id<Person> person = idDoubleEntry.getKey();
-            double emissionsSaved = idDoubleEntry.getValue() * 1.;
-            double klimaTaler = emissionsSaved / 1 * 1.0;
+            double emissionsSaved = idDoubleEntry.getValue() * 0.76;
+            double klimaTaler = emissionsSaved / 500 * 10;
             afterMobsimEvent.getServices().getEvents().processEvent(new PersonMoneyEvent(Time.MIDNIGHT, person, klimaTaler, "klimaTalerForPt", null, null));
         }
     }
@@ -134,11 +133,8 @@ public class KlimaTaler implements PersonDepartureEventHandler, PersonArrivalEve
         if (vehicles2Persons.containsKey(linkLeaveEvent.getVehicleId())) {
             Id<Person> personId = vehicles2Persons.get(linkLeaveEvent.getVehicleId());
             if (distanceTravelledBike.containsKey(personId)) {
-                log.info("bike leaves link " + linkLeaveEvent.getVehicleId());
                 double linkLength = network.getLinks().get(linkLeaveEvent.getLinkId()).getLength();
-                System.out.println("old distance: " + distanceTravelledBike.get(linkLeaveEvent.getVehicleId()));
                 double distanceTravelled = distanceTravelledBike.get(personId) + linkLength;
-                System.out.println("new distance: " + distanceTravelled);
                 distanceTravelledBike.replace(personId, distanceTravelled);
             }
